requirements.txt
jupyter==1.0.0
numpy==1.19.0
pandas==1.1.1
scikit-learn==0.23.1
seaborn==0.10.1
matplotlib==3.3.0
pillow==7.2.0
seaborn==0.10.1
tqdm==4.48.0
opencv-python==4.3.0.36
torch==1.7.0
torchvision==0.8.1
fastai==2.1.3
absl-py==0.10.0
kornia==0.4.1
scikit-image==0.17.2